+++ src/main/java/me/kall/enchantedpotato/common/config/MineCarveConfig.java
@@ -6,12 +6,15 @@ public class MineCarveConfig {
     public static final ModConfigSpec INSTANCE;
 
     public static final ModConfigSpec.DoubleValue BASE_ARMOR_REDUCTION, GAINED_ARMOR_REDUCTION_PER_LEVEL;
+    public static final ModConfigSpec.IntValue REDUCTION_TICKS_DURATION, GAINED_TICKS_DURATION_PER_LEVEL;
 
     static {
         ModConfigSpec.Builder builder = new ModConfigSpec.Builder();
         builder.push("MineCarve");
         BASE_ARMOR_REDUCTION = builder.defineInRange("BaseArmorReductionAmount", 2.0, 0.0, Double.MAX_VALUE);
         GAINED_ARMOR_REDUCTION_PER_LEVEL = builder.defineInRange("GainedArmorReductionAmountPerLevel", 0.5, 0.0, Double.MAX_VALUE);
+        REDUCTION_TICKS_DURATION = builder.defineInRange("ReductionTicksDuration", 600, 0, Integer.MAX_VALUE);
+        GAINED_TICKS_DURATION_PER_LEVEL = builder.defineInRange("GainedTicksDurationPerLevel", 300, 0, Integer.MAX_VALUE);
         builder.pop();
         INSTANCE = builder.build();
     }