+++ src/main/java/me/kall/enchantedpotato/common/enchantment/digger/MineCarve.java
@@ -1,5 +1,9 @@
 package me.kall.enchantedpotato.common.enchantment.digger;
 
+import it.unimi.dsi.fastutil.ints.IntObjectImmutablePair;
+import it.unimi.dsi.fastutil.ints.IntObjectPair;
+import it.unimi.dsi.fastutil.objects.Object2IntMap;
+import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import me.kall.enchantedpotato.EnchantedPotato;
 import me.kall.enchantedpotato.common.config.MineCarveConfig;
 import me.kall.enchantedpotato.common.config.json.DisableConfig;
@@ -18,10 +22,12 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.neoforged.neoforge.event.entity.living.LivingIncomingDamageEvent;
+import net.neoforged.neoforge.event.tick.ServerTickEvent;
 import org.jetbrains.annotations.NotNull;
 
 public class MineCarve extends BaseEnchantment {
     public static final BaseEnchantment INSTANCE = new MineCarve();
+    private static final Object2IntMap<IntObjectPair<Runnable>> TASKS = new Object2IntOpenHashMap<>();
 
     @Override
     public boolean isDisabled() {
@@ -76,6 +82,7 @@ public class MineCarve extends BaseEnchantment {
                 if (armor == null) return;
 
                 double amount = MineCarveConfig.BASE_ARMOR_REDUCTION.get() + MineCarveConfig.GAINED_ARMOR_REDUCTION_PER_LEVEL.get() * (double) (enchantmentLevel - 1);
+                int ticks = MineCarveConfig.REDUCTION_TICKS_DURATION.get() + MineCarveConfig.GAINED_TICKS_DURATION_PER_LEVEL.get() * (enchantmentLevel - 1);
 
                 AttributeModifier attributeModifier = new AttributeModifier(EnchantedPotato.loc("mine_carve_modifier"), -amount, AttributeModifier.Operation.ADD_VALUE);
 
@@ -84,8 +91,31 @@ public class MineCarve extends BaseEnchantment {
                 }
 
                 armor.addPermanentModifier(attributeModifier);
+
+                Runnable task = () -> {
+                    try {
+                        armor.removeModifier(attributeModifier);
+                    } catch (Throwable ignored) {}
+                    System.out.println("Removed modifier on " + event.getEntity());
+                };
+
+                TASKS.put(new IntObjectImmutablePair<>(ticks, task), 0);
             }
         }
     }
 
+    public static void onServerTick(ServerTickEvent.Post event) {
+        event.getServer().execute(() -> {
+            if (TASKS.isEmpty()) return;
+            new Object2IntOpenHashMap<>(TASKS).object2IntEntrySet().forEach(entry -> {
+                int ticks = entry.getIntValue() + 1;
+                if (ticks >= entry.getKey().firstInt()) {
+                    entry.getKey().second().run();
+                    TASKS.removeInt(entry.getKey());
+                } else {
+                    TASKS.put(entry.getKey(), ticks);
+                }
+            });
+        });
+    }
 }